% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled_data.R
\name{acled_data}
\alias{acled_data}
\title{ACLED API}
\usage{
acled_data(
  regions = "1|2|3|4|5",
  time.frame = NULL,
  more.variables = "",
  other.query = ""
)
}
\arguments{
\item{regions}{character string. Enter ACLED region numbers separated by |. Defaults to Africa: 1|2|3|4|5}

\item{time.frame}{character string. Enter the time period of observation to be retrieved. Format: "yyyy-mm-dd|yyyy-mm-dd"}

\item{more.variables}{character string. Enter names of ACLED variables you wish to add to the default output, each separated by |.}

\item{other.query}{character string. Enter other ACLED API queries you like to pass on to GET, with a & at the beginning of each term.}
}
\value{
A data frame object with ACLED events.
}
\description{
This package retrieves ACLED event data. It accesses the API of the Armed Conflict Location & Event Data Project (ACLED) at https://www.acleddata.com.
}
\author{
Christoph Dworschak
}
