% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.fun.R
\name{get.api.regions}
\alias{get.api.regions}
\title{ACLED API region support function}
\usage{
get.api.regions()
}
\value{
A list of length 2. \cr
\enumerate{
\item A data frame object containing ACLED region names and codes. \cr
\item A string with version information.
}
}
\description{
List of region names and codes as they are supplied by the \href{https://acleddata.com/resources/general-guides/}{ACLED API user guide}.
}
\details{
This single-purpose support function does not take arguments. It is set up as a convenient way to return region names and region codes
to be used in the \emph{\code{acled.api()}} main function's region argument.
}
\references{
Armed Conflict Location & Event Data Project (ACLED) \href{https://acleddata.com/resources/general-guides/}{API user guide} \cr
Clionadh Raleigh, Andrew Linke, Havard Hegre and Joakim Karlsen. 2010.
"Introducing ACLED-Armed Conflict Location and Event Data." \emph{Journal of Peace Research} 47 (5): 651-660.
}
\author{
Christoph Dworschak \cr Website: \href{https://www.chrisdworschak.com/}{\url{https://chrisdworschak.com/}}
}
