% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled.api.R
\name{acled.api}
\alias{acled.api}
\title{Automated Retrieval of ACLED Conflict Event Data}
\usage{
acled.api(
  regions = NULL,
  start.date = NULL,
  end.date = NULL,
  more.variables = NULL,
  all.variables = FALSE,
  dyadic = FALSE,
  other.query = NULL
)
}
\arguments{
\item{regions}{required numeric or character vector. Supply one or more region codes (numeric) or region names (character)
indicating the geographic region you wish to retrieve (see ACLED's codebook for details on region codes and names).}

\item{start.date}{required character string. Supply the earliest date to be retrieved. Format: "yyyy-mm-dd".}

\item{end.date}{required character string. Supply the last date to be retrieved. Format: "yyyy-mm-dd".}

\item{more.variables}{optional character vector. Supply the names of ACLED variables you wish to add to the
default output (see ACLED's codebook for details). Variables that are always are: region, country, year,
event_date, source, admin1, admin2, admin3, location, event_type, sub_event_type, interaction, fatalities.}

\item{all.variables}{optional logical. When set to FALSE (default), a default selection of variables is returned, which
can be refined using the argument more.variables). If set to TRUE, all variables are included in the output (overrides
argument more.variables).}

\item{dyadic}{optional logical. When set to FALSE (default), monadic data is returned (one
observation per event). If set to TRUE, dyadic data is returned.}

\item{other.query}{optional character vector. Allows users to add their own ACLED API queries to the
GET call. Note that some query terms require a ? in front.}
}
\value{
A data frame containing ACLED events.
}
\description{
This small package provides functionality to access and manage the application programming
interface (API) of the \href{https://acleddata.com/}{Armed Conflict Location & Event Data Project (ACLED)}.
The package \code{acleda.api} makes it easy to retrieve a user-defined sample (or all of the
available data) of ACLED, enabling a seamless integration of regular data updates into
the research work flow. \cr \cr
Important: When using this package, you acknowledge that you have read ACLED's terms and
conditions of use, and that you agree with their attribution requirements.
}
\details{
The package provides an R wrapper for the \href{https://acleddata.com/}{Armed Conflict Location & Event Data Project (ACLED)} API.
Internally it uses \code{httr} to access the API, and \code{jsonlite} to manage the JSON content that the call returns. The data
are converted into the base class \code{data.frame}. Variables are of class \code{character} by default. Variables which only contain numbers
as recognized by the regular expression \verb{^[0-9]+$} are coerced \code{as.numeric} before \code{data.frame} is returned.
}
\examples{
my.data.frame1 <- acled.api(regions = c(1,2,7),
start.date = "2018-11-01",
end.date = "2018-11-31")
head(my.data.frame1)

my.data.frame2 <- acled.api(regions = c(1,2,7),
start.date = "2018-11-01",
end.date = "2018-11-31",
more.variables = c("geo_precision", "time_precision"))
sd(my.data.frame2$geo_precision)
}
\references{
Armed Conflict Location & Event Data Project (ACLED); \url{https://acleddata.com/} \cr
Clionadh Raleigh, Andrew Linke, Havard Hegre and Joakim Karlsen. 2010.
"Introducing ACLED-Armed Conflict Location and Event Data." Journal of Peace Research 47 (5): 651-660.
}
\author{
Christoph Dworschak \cr Website: \href{https://www.chrisdworschak.com/}{\url{https://chrisdworschak.com/}}
}
