% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled.api.R
\name{acled.api}
\alias{acled.api}
\title{acled.api: Automated retrieval of ACLED conflict event data}
\usage{
acled.api(
  regions = NULL,
  start.date = NULL,
  end.date = NULL,
  more.variables = NULL,
  dyadic = NULL,
  other.query = NULL
)
}
\arguments{
\item{regions}{required numeric vector. Supply one or a vector of multiple integers indicating the geographic region
you wish to retrieve (see ACLED's codebook for details on region variable indicators).}

\item{start.date}{required character string. Supply the earliest date to be retrieved. Format: "yyyy-mm-dd".}

\item{end.date}{required character string. Supply the last date to be retrieved. Format: "yyyy-mm-dd".}

\item{more.variables}{optional character vector. Supply the names of ACLED variables you wish to add to the
default output (see ACLED's codebook for details). Variables that are always are: region, country, year,
event_date, source, admin1, admin2, admin3, location, event_type, sub_event_type, interaction, fatalities.}

\item{dyadic}{optional logical. When set to NULL (default) or FALSE, monadic data is returned (one
observation per event). If set to TRUE, dyadic data is returned.}

\item{other.query}{optional character vector. Allows users to add their own ACLED API queries to the
GET call. Note that some query terms require a ? in front.}
}
\value{
A data frame object containing ACLED events.
}
\description{
A small package to access the application programming interface (API) of
the Armed Conflict Location & Event Data
Project (ACLED) at \href{https://www.acleddata.com}{https://www.acleddata.com}. \cr \cr
When using this package, you acknowledge that you have read ACLED's terms and conditions
of use, and that you agree with their attribution requirements.
}
\examples{
my.data.frame1 <- acled.api(regions = c(1,2,7),
start.date = "2018-01-15",
end.date = "2018-12-31")
head(my.data.frame1)

my.data.frame2 <- acled.api(regions = c(1,2,7),
start.date = "2018-01-15",
end.date = "2018-12-31",
more.variables = c("geo_precision", "time_precision"))
sd(my.data.frame2$geo_precision)
}
\references{
Armed Conflict Location & Event Data Project (ACLED); https://www.acleddata.com \cr
Clionadh Raleigh, Andrew Linke, Havard Hegre and Joakim Karlsen. 2010.
"Introducing ACLED-Armed Conflict Location and Event Data." Journal of Peace Research 47 (5): 651-660.
}
\author{
Christoph Dworschak \cr Website: \href{https://www.chrisdworschak.com}{chrisdworschak.com}
}
